refactor exception to its own file
